domain 'entities' folder renamed 'models'
