+++ server/src/infrastructure/repositories/transaction.repository.ts
@@ -1,9 +1,9 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 
 import { ITransactionRepository } from '../../domain/repositories/transaction.repository';
-import { Transaction } from '../../domain/entities/transaction';
+import { Transaction } from '../../domain/models/transaction';
 import { CreateTransactionDTO } from '../controllers/transaction/transaction.create.dto';
 import {
   Transaction as TransactionModel,
